Initialized into Eclipse
