+++ .classpath
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<classpath>
+	<classpathentry kind="con" path="org.eclipse.jdt.launching.JRE_CONTAINER/org.eclipse.jdt.internal.debug.ui.launcher.StandardVMType/JavaSE-1.8"/>
+	<classpathentry kind="src" path="src"/>
+	<classpathentry kind="lib" path="E:/Testserver/spigot-1.12.2.jar"/>
+	<classpathentry kind="lib" path="C:/Users/Stijn/Downloads/Vault-1.6.7.jar"/>
+	<classpathentry kind="output" path="bin"/>
+</classpath>
